Исправление мелких замечаний по ревью
