+++ ImageFeedTests/WebViewTests.swift
@@ -8,22 +8,6 @@
 import XCTest
 @testable import ImageFeed
 
-
-final class WebViewPresenterSpy: WebViewPresenterProtocol {
-    var view: (any ImageFeed.WebViewViewControllerProtocol)?
-    var viewDidLoadCalled: Bool = false
-    
-    func viewDidLoad() {
-        viewDidLoadCalled = true
-    }
-    
-    func didUpdateProgressValue(_ newValue: Double) {}
-    
-    func code(from url: URL) -> String? {
-        return nil
-    }
-}
-
 final class WebViewTests: XCTestCase {
     
     func testViewControllerCallsViewDidLoad() {
